putting in two different pages, the Card Form and the CardList and adding the conditional rendering to the Welcome message
