+++ package.json
@@ -16,7 +16,8 @@
     "next": "13.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "styled-components": "^5.3.8"
+    "styled-components": "^5.3.8",
+    "use-local-storage-state": "^18.2.1"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
